+++ app/api/location/location.py
@@ -1,6 +1,8 @@
+from flask import request
 from flask_restful import Resource, reqparse, abort, fields, marshal_with
 from . import location_api
 from ..database_models import Location
+from ..decorator import token_required
 from .. import data_handler
 from app import db
 
@@ -59,6 +61,7 @@ resource_fields = {
 # Class to handle methods related to location
 class LocationResource(Resource):
     @marshal_with(resource_fields)
+    @token_required
     def get(self):
         try:
             args = location_get_args.parse_args()
@@ -70,6 +73,7 @@ class LocationResource(Resource):
         except Exception:
             abort(500, message="An internal server error has occurred, please try again later.")
 
+    @token_required
     def put(self):
         try:
             args = location_put_args.parse_args()
@@ -79,6 +83,7 @@ class LocationResource(Resource):
         except Exception:
             abort(500, message="An internal server error has occurred, please try again later.")
 
+    @token_required
     def patch(self):
         try:
             args = location_patch_args.parse_args()
@@ -88,12 +93,16 @@ class LocationResource(Resource):
         except Exception:
             abort(500, message="An internal server error has occurred, please try again later.")
 
+    @token_required
     def delete(self):
         try:
-            args = location_del_args.parse_args()
-            data_handler.clean_data(args)
-            delete_location(args)
-            return {"message": "Deleted from database"}, 204
+            if data_handler.check_if_admin(request.headers['x-access-token']):
+                args = location_del_args.parse_args()
+                data_handler.clean_data(args)
+                delete_location(args)
+                return {"message": "Deleted from database"}, 204
+            else:
+                abort(403, message="Forbidden.")
         except Exception:
             abort(500, message="An internal server error has occurred, please try again later.")
