+++ app/api/vaccination/vaccination.py
@@ -1,8 +1,10 @@
+from flask import request
 from flask_restful import Resource, reqparse, abort, fields, marshal_with
 from datetime import datetime
 import uuid
 from . import vaccination_api
 from ..database_models import Location, Vial, Vaccine, Citizen, Vaccination
+from ..decorator import token_required
 from .. import data_handler, sms_handler
 from app import db
 
@@ -78,6 +80,7 @@ resource_fields = {
 # where a citizen will be vaccinated
 class VaccinationResource(Resource):
     @marshal_with(resource_fields)
+    @token_required
     def get(self):
         try:
             args = vaccination_get_args.parse_args()
@@ -86,6 +89,7 @@ class VaccinationResource(Resource):
         except Exception:
             abort(500, message="An internal server error has occurred, please try again later.")
 
+    @token_required
     def put(self):
         try:
             args = vaccination_put_args.parse_args()
@@ -95,6 +99,7 @@ class VaccinationResource(Resource):
         except Exception:
             abort(500, message="An internal server error has occurred, please try again later.")
 
+    @token_required
     def patch(self):
         try:
             args = vaccination_patch_args.parse_args()
@@ -104,12 +109,16 @@ class VaccinationResource(Resource):
         except Exception:
             abort(500, message="An internal server error has occurred, please try again later.")
 
+    @token_required
     def delete(self):
         try:
-            args = vaccination_del_args.parse_args()
-            data_handler.remove_space(args)
-            delete_vaccination(args)
-            return {"message": "Deleted from database"}, 204
+            if data_handler.check_if_admin(request.headers['x-access-token']):
+                args = vaccination_del_args.parse_args()
+                data_handler.remove_space(args)
+                delete_vaccination(args)
+                return {"message": "Deleted from database"}, 204
+            else:
+                abort(403, message="Forbidden.")
         except Exception:
             abort(500, message="An internal server error has occurred, please try again later.")
